SMARTAppLaunch

SMARTAppLaunch.ProviderEHRLaunchConfig

SMARTAppLaunch.ProviderStandaloneConfig

SMARTAppLaunch.get_access_token

SMARTAppLaunch.get_access_token

SMARTAppLaunch.provider_ehr_launch

SMARTAppLaunch.provider_ehr_launch

SMARTAppLaunch.provider_standalone_step1

SMARTAppLaunch.provider_standalone_step2

SMARTAppLaunch.provider_standalone_step2

SMARTAppLaunch.ProviderEHRLaunchConfig — Type

ProviderEHRLaunchConfig(; kwargs...)

Required Keyword Arguments:

client_id::String

redirect_uri::String

SMARTAppLaunch.ProviderStandaloneConfig — Type

ProviderStandaloneConfig(; kwargs...)

Required Keyword Arguments:

authorize_endpoint::String

client_id::String

redirect_uri::String

token_endpoint::String

SMARTAppLaunch.get_access_token — Method

get_access_token(result::ProviderEHRLaunchResult)

SMARTAppLaunch.get_access_token — Method

get_access_token(result::ProviderStandaloneResult)

SMARTAppLaunch.provider_ehr_launch — Method

provider_ehr_launch(config::ProviderEHRLaunchConfig, uri_string::AbstractString; kwargs...)

Optional Keyword Arguments:

scope::AbstractString. Default value: "launch".

SMARTAppLaunch.provider_ehr_launch — Method

provider_ehr_launch(config::ProviderEHRLaunchConfig, uri::URIs.URI; kwargs...)

Optional Keyword Arguments:

scope::AbstractString. Default value: "launch".

SMARTAppLaunch.provider_standalone_step1 — Method

provider_standalone_step1(config::ProviderStandaloneConfig; kwargs...)

Optional Keyword Arguments:

scope::AbstractString. Default value: "launch".

SMARTAppLaunch.provider_standalone_step2 — Method

provider_standalone_step2(config::ProviderStandaloneConfig, uri_string::AbstractString)

SMARTAppLaunch.provider_standalone_step2 — Method

provider_standalone_step2(config::ProviderStandaloneConfig, uri::URIs.URI)

This package implements the SMART App Launch Framework for building SMART-on-FHIR applications.

The source code for this package is available in the GitHub repository.

The following tables show the mapping between Julia packages and standards/specifications:

Julia Package	Standard/Specification	Description

FHIRClient.jl	FHIR	Fast Healthcare Interoperability Resources. Web standard for health interop.

SMARTAppLaunch.jl	SMART App Launch	User-facing apps that connect to EHRs and health portals.

We currently do not implement the following; however, we plan to implement them in the future:

Standard/Specification	Description

CDS Hooks	Clinical Decision Support Hooks. Web standard for CDS in the EHR workflow.

FHIR Bulk Data API	FHIR export API for large-scale data access.

SMART Backend Services	Server-to-server FHIR connections.

These descriptions are taken from https://docs.smarthealthit.org/.